refactor: Standardize formatter abstractions to use `IFormatterDefinition` and update generic type parameters
This refactoring commit standardizes the naming of formatter-related interfaces and their generic type parameters to improve consistency across the codebase. The changes align the formatter system with a broader framework convention where "Definition" is used for types that provide semantic metadata.

The key changes include:

-   **Interface and File Renaming**: The `IFormatterType` interface is renamed to **`IFormatterDefinition`**, and its file is moved to `src/Formatters/Definitions/IFormatterDefinition.cs`. This clarifies that the interface describes the metadata of a formatter rather than its runtime type.

-   **Generic Type Parameter Renaming**: The generic type parameters in `IFormatter` and `IFormatterFactory` are updated to improve clarity.
    -   `IFormatter<TIn, out TOut>` is changed to `IFormatter<in TIn, TOut>`. The `out` keyword for the `TOut` type parameter is removed as it's not strictly necessary for this interface's contract and can lead to confusion. This is a subtle but important change to the generic constraint.
    -   `IFormatterFactory.cs` is renamed to `IFormatterFactory{TIn, TOut}.cs` to reflect its generic nature.

-   **Reference Updates**: All references to `IFormatterType` are updated to point to the new `IFormatterDefinition` interface.
    -   `IFormatterFactory` now uses `IFormatterDefinition` as a parameter type.
    -   `IFormatter`'s `Type` property is renamed to **`Definition`** to be more semantically correct, and its type is updated to `IFormatterDefinition`.
    -   `IFormatterRegistry` is updated to return and handle `IFormatterDefinition` instances.

This refactoring is a non-functional change that enhances the readability and maintainability of the code by establishing a consistent and more explicit naming standard.
