+++ src/Formatters/IFormatter{in TIn, TOut}.cs
@@ -1,8 +1,9 @@
-// <copyright file="IFormatter{in TIn, out TOut}.cs" company="Zentient Framework Team">
+// <copyright file="IFormatter{in TIn, TOut}.cs" company="Zentient Framework Team">
 // Copyright © 2025 Zentient Framework Team. All rights reserved.
 // </copyright>
 
 using Zentient.Abstractions.Common;
+using Zentient.Abstractions.Formatters.Definitions;
 
 namespace Zentient.Abstractions.Formatters
 {
@@ -11,12 +12,12 @@ namespace Zentient.Abstractions.Formatters
     /// </summary>
     /// <typeparam name="TIn">The type of input value to be formatted.</typeparam>
     /// <typeparam name="TOut">The type of output value after formatting.</typeparam>
-    public interface IFormatter<TIn, TOut>
+    public interface IFormatter<in TIn, TOut>
     {
         /// <summary>
         /// Gets the formatter's type metadata, including identity, name, and semantic info.
         /// </summary>
-        IFormatterType Type { get; }
+        IFormatterDefinition definition { get; }
 
         /// <summary>Transforms the input value asynchronously.</summary>
         /// <param name="input">The input value.</param>
